Migrated cart model to it's own file
