+++ server/db/index.js
@@ -5,23 +5,10 @@ const User = require('./models/User');
 const Liquor = require('./models/Liquor');
 const Cart = require('./models/Cart');
 const Sequelize = require('sequelize');
+const cartLiquor = require('./models/cartLiquors');
 //associations could go here!
 
-const cartLiquor = db.define('cartLiquor', {
-  liquorQuantity: {
-    type: Sequelize.INTEGER,
-    defaultValue: 0,
-  },
-  liquorPrice: {
-    type: Sequelize.FLOAT,
-    defaultValue: 0,
-  },
-  image: {
-    type: Sequelize.STRING,
-    defaultValue:
-      'https://packagingoptionsdirect.com/750-ml-clear-glass-oval-liquor-bottle-215-mm-bar-top-neck-finish',
-  },
-});
+
 
 Cart.belongsTo(User);
 User.hasOne(Cart);
